most features working
A Python script aftercare.py adds the import extensions as before as well as empty ownDefaults and ownMutabilities methods to the classes that miss them. This fixes a bug in the assignment of default values.

As a bonus, defaults and mutabilities no longer require calls to custom updateDefaults and updateMutabilities methods.